基本完成活动模块，初步完成小组模块

+++ src/main/java/com/xdu/formteamtalent/controller/UserController.java
@@ -4,6 +4,8 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import com.xdu.formteamtalent.entity.*;
 import com.xdu.formteamtalent.global.RestfulResponse;
+import com.xdu.formteamtalent.service.ActivityService;
+import com.xdu.formteamtalent.service.TeamService;
 import com.xdu.formteamtalent.service.UserService;
 import com.xdu.formteamtalent.service.UATService;
 import com.xdu.formteamtalent.utils.JwtUtil;
@@ -19,6 +21,18 @@ import javax.servlet.http.HttpServletResponse;
 public class UserController {
     private UserService userService;
     private UATService uatService;
+    private ActivityService activityService;
+    private TeamService teamService;
+
+    @Autowired
+    public void setTeamService(TeamService teamService) {
+        this.teamService = teamService;
+    }
+
+    @Autowired
+    public void setActivityService(ActivityService activityService) {
+        this.activityService = activityService;
+    }
 
     @Autowired
     public void setUatService(UATService uatService) {
@@ -71,15 +85,52 @@ public class UserController {
     }
 
     /**
+     * 加入某个活动
+     * @param a_id 活动id
+     */
+    @PostMapping("/join/activity")
+    public RestfulResponse joinActivity(HttpServletRequest request, @RequestParam("a_id") String a_id) {
+        UAT uat = new UAT();
+        uat.setU_id(WxUtil.getOpenId(request));
+        uat.setA_id(a_id);
+        uatService.save(uat);
+        return RestfulResponse.success();
+    }
+
+    /**
      * 退出某个小组
      * @param t_id 小组id
      */
     @PostMapping("/leave/team")
 
     public RestfulResponse leaveTeam(HttpServletRequest request, @RequestParam("t_id") String t_id) {
+        String u_id = WxUtil.getOpenId(request);
         QueryWrapper<UAT> wrapper = new QueryWrapper<>();
         wrapper.eq("t_id", t_id);
-        wrapper.eq("u_id", WxUtil.getOpenId(request));
+        wrapper.eq("u_id", u_id);
+        Team team = teamService.getOne(new QueryWrapper<Team>().eq("t_leader_id", u_id));
+        if (team != null) {
+            return RestfulResponse.fail(403, "作为小组创建者，需要先解散该队伍");
+        }
+        uatService.remove(wrapper);
+        return RestfulResponse.success();
+    }
+
+    /**
+     * 退出某个活动
+     * @param a_id 活动id
+     */
+    @PostMapping("/leave/activity")
+
+    public RestfulResponse leaveActivity(HttpServletRequest request, @RequestParam("a_id") String a_id) {
+        String u_id = WxUtil.getOpenId(request);
+        QueryWrapper<UAT> wrapper = new QueryWrapper<>();
+        wrapper.eq("a_id", a_id);
+        wrapper.eq("u_id", u_id);
+        Activity activity = activityService.getOne(new QueryWrapper<Activity>().eq("a_holder_id", u_id));
+        if (activity != null) {
+            return RestfulResponse.fail(403, "作为活动创建者，需先删除该活动");
+        }
         uatService.remove(wrapper);
         return RestfulResponse.success();
     }
